+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this application adheres to [Semantic Versioning](https://semver.org/spec/v2
 
 ### Changed
 - The About screen has been redesigned to align with the new design guidelines
+- `StyledBalance` text styles have been refactored from `Pair` into `BalanceTextStyle` 
 
 ## [1.1.3 (682)] - 2024-07-03